minimal working example of disk image analysis
