+++ bin/main.go
@@ -74,6 +74,15 @@ var (
 	tempdir_flag = app.Flag(
 		"tempdir", "Write all temp files to this directory").String()
 
+	device_flag = app.Flag(
+		"device", "Use this device as the analysis target instead of the local filesystem.").String()
+
+	device_accessor_flag = app.Flag(
+		"device_accessor", "The accessor type required to open `device'.").String()
+
+	analysis_target_flag = app.Flag(
+		"analysis_target", "The OS of the analysis subject.").String()
+
 	command_handlers []CommandHandler
 )
 
@@ -155,6 +164,7 @@ func main() {
 	// - first try the API config, then try a config.
 	APIConfigLoader = new(config.Loader).WithVerbose(*verbose_flag).
 		WithTempdir(*tempdir_flag).
+		WithCustomValidator(applyDiskAnalysisSettings).
 		WithApiLoader(*api_config_path).
 		WithEnvApiLoader("VELOCIRAPTOR_API_CONFIG").
 		WithCustomValidator(maybe_unlock_api_config).
@@ -198,6 +208,7 @@ func makeDefaultConfigLoader() *config.Loader {
 	return new(config.Loader).
 		WithVerbose(*verbose_flag).
 		WithTempdir(*tempdir_flag).
+		WithCustomValidator(applyDiskAnalysisSettings).
 		WithFileLoader(*config_path).
 		WithEmbedded().
 		WithEnvLoader("VELOCIRAPTOR_CONFIG").
