+++ vql/windows/filesystems/ntfs_cache.go
@@ -1,5 +1,3 @@
-// +build windows
-
 package filesystems
 
 import (
