+++ vql_plugins/plugins_linux.go
@@ -19,4 +19,5 @@ package plugins
 
 import (
 	_ "www.velocidex.com/golang/velociraptor/vql/linux"
+	_ "www.velocidex.com/golang/velociraptor/vql/windows/filesystems"
 )
